+++ Dyalect/Compiler/OpSizeHelper.cs
@@ -90,7 +90,6 @@
             0, //Mut
             0, //Annot
             1, //FunAttr
-            1, //NewErr
             0, //NewLaz
             0, //NewCast
             0, //Cast