+++ Dyalect/DyType.cs
@@ -20,8 +20,8 @@ namespace Dyalect
         public const int Tuple = 13;
         public const int Dictionary = 14;
         public const int Set = 15;
-        public const int Error = 16;
-        public const int Lazy = 17;
+        public const int Lazy = 16;
+        public const int Variant = 17;
         public const int Object = 18;
 
         internal static List<DyTypeInfo> GetAll() =>
@@ -43,8 +43,8 @@ namespace Dyalect
                 new DyTupleTypeInfo(),
                 new DyDictionaryTypeInfo(),
                 new DySetTypeInfo(),
-                new DyErrorTypeInfo(),
                 new DyLazyTypeInfo(),
+                new DyVariantTypeInfo(),
                 new DyObjectTypeInfo()
             };
 
@@ -66,8 +66,8 @@ namespace Dyalect
                 DyTypeNames.Iterator => Iterator,
                 DyTypeNames.Dictionary => Dictionary,
                 DyTypeNames.Set => Set,
-                DyTypeNames.Error => Error,
                 //DyTypeNames.Lazy => Lazy,
+                DyTypeNames.Variant => Variant,
                 DyTypeNames.Object => Object,
                 _ => default
             }; 
@@ -90,8 +90,8 @@ namespace Dyalect
                 Iterator => DyTypeNames.Iterator,
                 Dictionary => DyTypeNames.Dictionary,
                 Set => DyTypeNames.Set,
-                Error => DyTypeNames.Error,
                 Lazy => DyTypeNames.Lazy,
+                Variant => DyTypeNames.Variant,
                 Object => DyTypeNames.Object,
                 _ => code.ToString(),
             };
@@ -117,8 +117,8 @@ namespace Dyalect
                 Tuple,
                 Dictionary,
                 Set,
-                Error,
                 Lazy,
+                Variant,
                 Object
             };
 
@@ -137,8 +137,8 @@ namespace Dyalect
         public const string Tuple = "Tuple";
         public const string Dictionary = "Dictionary";
         public const string Set = "Set";
-        public const string Error = "Error";
         public const string Lazy = "Lazy";
+        public const string Variant = "Variant";
         public const string Object = "Object";
     }
 }